mount the code as volume to avoid rebuilding the container every time
