+++ leeloo/Dockerfile
@@ -9,7 +9,7 @@ RUN pip install -r requirements.txt
 RUN git clone https://github.com/uktrade/data-hub-backend.git
 RUN pip install -e data-hub-backend/korben
 
-ADD . /app/leeloo
+ADD . /app/leeloo/
 
 RUN chmod a+x start.sh
