--- leeloo/api/test/test_companies.py
@@ -1,5 +1,0 @@
-"""Tests"""
-
-def test_company_get():
-    """Testing test (pun not intended), to be deleted soon."""
-    assert True
