--- leeloo/gunicorn/conf.py
@@ -1,4 +1,0 @@
-import gunicorn
-
-
-gunicorn.SERVER_SOFTWARE = "Wonderland/1.0"
