+++ leeloo/search/services.py
@@ -1,0 +1,1 @@
+from korben.client import Client
